Change timestamp data type, from uint256 to uint64

+++ ethereum/contracts/atomic.sol
@@ -17,7 +17,7 @@ contract AtomicSwapper {
 
         uint256 expireHeight;
         bytes32 secretKey;
-        uint256 timestamp;
+        uint64  timestamp;
 
         address sender;
         address receiverAddr;
@@ -32,7 +32,7 @@ contract AtomicSwapper {
     }
 
     // Events
-    event SwapInitialization(address indexed _msgSender, address indexed _receiverAddr, bytes20 _BEP2Addr, uint256 _index, bytes32 _secretHashLock, uint256 _timestamp, uint256 _expireHeight, uint256 _erc20Amount, uint256 _bep2Amount);
+    event SwapInitialization(address indexed _msgSender, address indexed _receiverAddr, bytes20 _BEP2Addr, uint256 _index, bytes32 _secretHashLock, uint64 _timestamp, uint256 _expireHeight, uint256 _erc20Amount, uint256 _bep2Amount);
     event SwapExpire(address indexed _msgSender, address indexed _swapSender, bytes32 _secretHashLock);
     event SwapCompletion(address indexed _msgSender, address indexed _receiverAddr, bytes32 _secretHashLock, bytes32 _secretKey);
 
@@ -93,7 +93,7 @@ contract AtomicSwapper {
     /// @param _bep2Amount BEP2 asset to swap in.
     function initiate(
         bytes32 _secretHashLock,
-        uint256 _timestamp,
+        uint64  _timestamp,
         uint256 _timelock,
         address _receiverAddr,
         bytes20 _BEP2Addr,
@@ -161,7 +161,7 @@ contract AtomicSwapper {
     /// @notice query an atomic swap by secretHashLock
     ///
     /// @param _secretHashLock The hash of secretKey and timestamp
-    function querySwapByHashLock(bytes32 _secretHashLock) external view returns(uint256 _timestamp, uint256 _expireHeight, uint256 _erc20Amount, uint256 _bep2Amount, address _sender, address _receiver, bytes20 _BEP2Addr, bytes32 _secretKey, States _status) {
+    function querySwapByHashLock(bytes32 _secretHashLock) external view returns(uint64 _timestamp, uint256 _expireHeight, uint256 _erc20Amount, uint256 _bep2Amount, address _sender, address _receiver, bytes20 _BEP2Addr, bytes32 _secretKey, States _status) {
         Swap memory swap = swaps[_secretHashLock];
         States status = swapStates[_secretHashLock];
         return (
@@ -180,7 +180,7 @@ contract AtomicSwapper {
     /// @notice query an atomic swap by swap index
     ///
     /// @param _index The swap index
-    function querySwapByIndex(uint256 _index) external view returns (bytes32 _secretHashLock, uint256 _timestamp, uint256 _expireHeight, uint256 _erc20Amount, uint256 _bep2Amount, address _sender, address _receiver, bytes20 _BEP2Addr, bytes32 _secretKey, States _status) {
+    function querySwapByIndex(uint256 _index) external view returns (bytes32 _secretHashLock, uint64 _timestamp, uint256 _expireHeight, uint256 _erc20Amount, uint256 _bep2Amount, address _sender, address _receiver, bytes20 _BEP2Addr, bytes32 _secretKey, States _status) {
         bytes32 secretHashLock = indexToSecretHashLock[_index];
         Swap memory swap = swaps[secretHashLock];
         States status = swapStates[secretHashLock];
@@ -224,7 +224,7 @@ contract AtomicSwapper {
     ///
     /// @param _secretKey The secret.
     /// @param _timestamp The timestamp.
-    function calSecretHash(bytes32 _secretKey, uint256 _timestamp) public pure returns (bytes32) {
+    function calSecretHash(bytes32 _secretKey, uint64 _timestamp) public pure returns (bytes32) {
         return sha256(abi.encodePacked(_secretKey, _timestamp));
     }
 }
